requirements.txt
streamlit
googletrans-py
gtts
PyPDF2
requests
